Improved File structure
